third_party/blink/web_tests/external/wpt/fs/FileSystemObserver-writable-file-stream.https.tentative.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] Closing a FileSystemWritableFileStream that's modified the file produces a "modified" event
  assert_equals: A record's type didn't match the expected type expected "modified" but got "appeared"
Harness: the test ran to completion.
